+++ cmd/root.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const Version = "2.3.0"
+const Version = "2.4.0"
 
 var rootCmd = &cobra.Command{
 	Use:   "snare",
@@ -45,4 +45,5 @@ func init() {
 	rootCmd.AddCommand(sessionCmd)
 	rootCmd.AddCommand(testCmd)
 	rootCmd.AddCommand(fuzzCmd)
+	rootCmd.AddCommand(bundleCmd)
 }